[XPU] Reorganize xpu device codes in platform, test=develop
