+++ paddle/fluid/framework/details/bkcl_op_handle.h
@@ -23,7 +23,7 @@
 #include "paddle/fluid/framework/details/op_handle_base.h"
 #include "paddle/fluid/framework/lod_tensor.h"
 #include "paddle/fluid/framework/scope.h"
-#include "paddle/fluid/platform/bkcl_helper.h"
+#include "paddle/fluid/platform/device/xpu/bkcl_helper.h"
 
 DECLARE_bool(sync_bkcl_allreduce);
