+++ paddle/fluid/imperative/bkcl_context.cc
@@ -20,8 +20,8 @@
 #include <vector>
 
 #include "paddle/fluid/framework/variable.h"
-#include "paddle/fluid/platform/bkcl_helper.h"
 #include "paddle/fluid/platform/collective_helper.h"
+#include "paddle/fluid/platform/device/xpu/bkcl_helper.h"
 #include "paddle/fluid/platform/device_context.h"
 #include "paddle/fluid/platform/gen_comm_id_helper.h"
 #include "paddle/fluid/platform/place.h"
