+++ paddle/fluid/memory/allocation/allocator_facade.cc
@@ -42,7 +42,7 @@
 #include "paddle/fluid/platform/cuda_graph.h"
 #endif
 #ifdef PADDLE_WITH_XPU
-#include "paddle/fluid/platform/xpu/xpu_info.h"
+#include "paddle/fluid/platform/device/xpu/xpu_info.h"
 #endif
 #include "paddle/fluid/platform/npu_info.h"
