+++ paddle/fluid/memory/allocation/naive_best_fit_allocator.cc
@@ -31,7 +31,7 @@
 #include "paddle/fluid/platform/cuda_device_guard.h"
 #endif
 #ifdef PADDLE_WITH_XPU
-#include "paddle/fluid/platform/xpu/xpu_header.h"
+#include "paddle/fluid/platform/device/xpu/xpu_header.h"
 #endif
 
 PADDLE_DEFINE_EXPORTED_bool(
