+++ paddle/fluid/operators/reduce_ops/logsumexp_op_xpu.cc
@@ -15,8 +15,8 @@
 #ifdef PADDLE_WITH_XPU
 
 #include "paddle/fluid/operators/reduce_ops/logsumexp_op.h"
+#include "paddle/fluid/platform/device/xpu/xpu_header.h"
 #include "paddle/fluid/platform/device_context.h"
-#include "paddle/fluid/platform/xpu/xpu_header.h"
 
 namespace paddle {
 namespace operators {
