+++ paddle/fluid/operators/sum_op_xpu.cc
@@ -13,7 +13,7 @@ limitations under the License. */
 
 #include "paddle/fluid/operators/sum_op.h"
 #include <vector>
-#include "paddle/fluid/platform/xpu/xpu_header.h"
+#include "paddle/fluid/platform/device/xpu/xpu_header.h"
 
 namespace paddle {
 namespace operators {
