+++ paddle/fluid/platform/CMakeLists.txt
@@ -4,12 +4,6 @@ if(WITH_GPU)
   proto_library(external_error_proto SRCS external_error.proto)
 endif(WITH_GPU)
 
-if(WITH_XPU)
-  set(XPU_CTX_DEPS xpulib ssl crypto rt z resolv dl)
-ELSE()
-  set(XPU_CTX_DEPS)
-endif(WITH_XPU)
-
 if(WITH_ASCEND)
     set(ASCEND_DEPS xpulib)
 ELSE()
@@ -74,10 +68,7 @@ ENDIF()
 cc_library(place SRCS place.cc DEPS enforce boost)
 cc_test(place_test SRCS place_test.cc DEPS place glog gflags)
 
-if(WITH_XPU)
-cc_library(xpu_info SRCS xpu/xpu_info.cc DEPS gflags glog enforce xpulib)
-cc_library(xpu_op_list SRCS xpu/xpu_op_list.cc DEPS gflags glog enforce xpulib device_context)
-endif()
+add_subdirectory(device)
 
 if(WITH_ASCEND)
     cc_library(ascend_npu_info SRCS ascend_npu_info.cc DEPS gflags glog enforce atlas_acl)
