+++ paddle/fluid/platform/device/CMakeLists.txt
@@ -1,0 +1,4 @@
+# XPU
+IF(WITH_XPU)
+  add_subdirectory(xpu)
+ENDIF()
