+++ paddle/fluid/platform/device/xpu/xpu2_op_list.h
