+++ paddle/fluid/platform/device/xpu/xpu_header.h
