+++ paddle/fluid/platform/device/xpu/xpu_info.h
